Type overhaul. More tracing. Lookup fix.
Concepts:
TypeName is an alias for String.
ValueType has a name() of TypeName.
MethodType has a ret() of ValueType and args() of Vector<ValueType>.
Type is either a ValueType or a MethodType.

TypeChecker functions have been reworked to take/return the most
specific type possible (i.e. ValueType over Type or TypeName when that's
the only thing allowed).

Added --ptrace and --ttrace for only tracing the parser or typechecker.

lookup for Map and Set return pointers now instead of Optionals.